Split psycopg2 vs sqlalchemy code
