+++ itou/metabase/management/commands/_database_psycopg2.py
@@ -1,0 +1,20 @@
+import psycopg2
+from django.conf import settings
+
+
+class MetabaseDatabaseCursor:
+    def __enter__(self):
+        self.conn = psycopg2.connect(
+            host=settings.METABASE_HOST,
+            port=settings.METABASE_PORT,
+            dbname=settings.METABASE_DATABASE,
+            user=settings.METABASE_USER,
+            password=settings.METABASE_PASSWORD,
+        )
+        self.cur = self.conn.cursor()
+        return self.cur, self.conn
+
+    def __exit__(self, exc_type, exc_value, exc_traceback):
+        self.conn.commit()
+        self.cur.close()
+        self.conn.close()
