+++ itou/metabase/management/commands/_database_sqlalchemy.py
@@ -1,4 +1,3 @@
-import psycopg2
 from django.conf import settings
 from sqlalchemy import create_engine
 
@@ -22,21 +21,3 @@ PG_ENGINE = create_engine(
     f"postgresql://{settings.METABASE_USER}:{settings.METABASE_PASSWORD}"
     f"@{settings.METABASE_HOST}:{settings.METABASE_PORT}/{settings.METABASE_DATABASE}"
 )
-
-
-class MetabaseDatabaseCursor:
-    def __enter__(self):
-        self.conn = psycopg2.connect(
-            host=settings.METABASE_HOST,
-            port=settings.METABASE_PORT,
-            dbname=settings.METABASE_DATABASE,
-            user=settings.METABASE_USER,
-            password=settings.METABASE_PASSWORD,
-        )
-        self.cur = self.conn.cursor()
-        return self.cur, self.conn
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        self.conn.commit()
-        self.cur.close()
-        self.conn.close()
